FootbolTV
This code enables voting in a football match where anyone can choose which player to vote for. However, there are only 23 players available, and if the entered number is 0, the voting ends. If the number is greater than 23, it prompts to repeat the input as there are no players beyond 23."